move evm address script into jellyfish-address
