+++ packages/jellyfish-crypto/package.json
@@ -14,7 +14,6 @@
     "build": "tsc -b ./tsconfig.build.json"
   },
   "dependencies": {
-    "@defichain/jellyfish-transaction": "^0.0.0",
     "bech32": "^2.0.0",
     "bip66": "^1.1.5",
     "browserify-aes": "^1.2.0",
