+++ packages/jellyfish-crypto/src/eth.ts
@@ -1,6 +1,5 @@
 import createKeccakHash from 'keccak'
 import { KECCAK256 } from './hash'
-import { OP_CODES, Script } from '@defichain/jellyfish-transaction'
 
 // https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md
 function toChecksumAddress (address: string): string {
@@ -24,12 +23,6 @@ function toAddress (hash: Buffer): string {
   return `0x${sliced}`
 }
 
-function validateAddress (address: string): boolean {
-  // https://github.com/ethers-io/ethers.js/blob/5210b68a7837654c6b84207a45e1e573d9472d1a/src.ts/address/address.ts#L123
-  const regex: RegExp = /^0x[a-fA-F0-9]{40}$/gm
-  return regex.test(address)
-}
-
 export const Eth = {
   /**
    * @param {Buffer} uncompressed pubKey to format into Eth
@@ -42,20 +35,5 @@ export const Eth = {
     const sub = pubKeyUncompressed.subarray(1, 65)
     const hash = KECCAK256(sub)
     return toChecksumAddress(toAddress(hash))
-  },
-  /**
-   * @param {string} address to convert into Script
-   * @return {Script} Script parsed from address
-   */
-  fromAddress (address: string): Script | undefined {
-    if (!validateAddress(address)) {
-      return undefined
-    }
-    return {
-      stack: [
-        OP_CODES.OP_16,
-        OP_CODES.OP_PUSHDATA_HEX_LE(address.substring(2))
-      ]
-    }
   }
 }
